requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
PyPDF2==3.0.1
python-docx==1.0.1
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.2
python-jose==3.3.0 